QRgenerator/requirements.txt
Pillow==11.0.0
PyQRCode==1.2.1
